ai_needle_emg-master/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
audioread==2.1.9
backports.functools-lru-cache==1.6.1
certifi==2021.5.30
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
et-xmlfile==1.0.1
ffmpeg-normalize==1.20.2
flatbuffers==1.12
gast==0.3.3
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.0.1
importlib-metadata==4.0.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.17.2
joblib==1.0.0
jupyter==1.0.0
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.5
jupyter-nbextensions-configurator==0.4.1
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.0
lxml==4.6.3
Markdown==3.3.3
matplotlib==3.3.4
mistune==0.8.3
nbclient==0.5.2
numba==0.53.1
numpy==1.21.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.6
opt-einsum==3.3.0
pandas==1.2.1
parso==0.7.0
path==15.1.2
Pillow==8.1.0
pip==21.1.3
pooch==1.3.0
prompt-toolkit==3.0.17
protobuf==3.14.0
#PyAudio==0.2.11
pycparser==2.19
Pygments==2.8.1
pyloudnorm==0.1.0
pyparsing==2.4.7
#PyQt5==5.12.3
#PyQt5-sip==4.19.18
#PyQtChart==5.12
#PyQtWebEngine==5.12.1
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
qtconsole==5.0.3
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.6.0
setuptools==49.6.0
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tk==0.1.0
tornado==6.1
tqdm==4.56.0
Wave==0.0.2
Werkzeug==1.0.1
wheel==0.36.2
wincertstore==0.2
xlrd==2.0.1
zmq==0.0.0